[WIFI-11251] Now fetching device statistics in batches of 100
