+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ucentral-client",
-  "version": "2.8.0(17)",
+  "version": "2.8.0(18)",
   "description": "",
   "private": true,
   "main": "index.tsx",
