+++ src/pages/Device/StatisticsCard/index.tsx
@@ -5,25 +5,13 @@ import { v4 as uuid } from 'uuid';
 import StatisticsCardDatePickers from './DatePickers';
 import InterfaceChart from './InterfaceChart';
 import DeviceMemoryChart from './MemoryChart';
+import { useStatisticsCard } from './useStatisticsCard';
 import ViewLastStatsModal from './ViewLastStatsModal';
 import { RefreshButton } from 'components/Buttons/RefreshButton';
 import { Card } from 'components/Containers/Card';
 import { CardBody } from 'components/Containers/Card/CardBody';
 import { CardHeader } from 'components/Containers/Card/CardHeader';
 import { LoadingOverlay } from 'components/LoadingOverlay';
-import {
-  DeviceStatistics,
-  useGetDeviceStatsLatestHour,
-  useGetDeviceStatsWithTimestamps,
-} from 'hooks/Network/Statistics';
-
-const extractMemory = (stat: DeviceStatistics) => {
-  let used: number | undefined;
-  if (stat.unit && stat.unit.memory) {
-    used = stat.unit.memory.total - stat.unit.memory.free;
-  }
-  return { ...stat.unit?.memory, used };
-};
 
 type Props = {
   serialNumber: string;
@@ -31,19 +19,9 @@ type Props = {
 
 const DeviceStatisticsCard = ({ serialNumber }: Props) => {
   const { t } = useTranslation();
-  const [time, setTime] = React.useState<{ start: Date; end: Date } | undefined>();
-  const [selected, setSelected] = React.useState('memory');
-  const getStats = useGetDeviceStatsLatestHour({ serialNumber, limit: 10000 });
-  const getCustomStats = useGetDeviceStatsWithTimestamps({
+  const { time, setTime, parsedData, isLoading, selected, onSelectInterface, refresh } = useStatisticsCard({
     serialNumber,
-    limit: 10000,
-    start: time ? Math.floor(time.start.getTime() / 1000) : undefined,
-    end: time ? Math.floor(time.end.getTime() / 1000) : undefined,
   });
-
-  const onChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    setSelected(event.target.value);
-  };
 
   const setNewTime = (start: Date, end: Date) => {
     setTime({ start, end });
@@ -51,79 +29,6 @@ const DeviceStatisticsCard = ({ serialNumber }: Props) => {
   const onClear = () => {
     setTime(undefined);
   };
-
-  const parsedData = React.useMemo(() => {
-    if (!getStats.data && !getCustomStats.data) return undefined;
-
-    const data: Record<string, { tx: number[]; rx: number[]; recorded: number[]; maxRx: number; maxTx: number }> = {};
-    const memoryData = {
-      used: [] as number[],
-      buffered: [] as number[],
-      cached: [] as number[],
-      free: [] as number[],
-      total: [] as number[],
-      recorded: [] as number[],
-    };
-    const previousRx: { [key: string]: number } = {};
-    const previousTx: { [key: string]: number } = {};
-
-    const dataToLoop = getCustomStats.data ?? getStats.data;
-
-    for (const [index, stat] of dataToLoop ? dataToLoop.data.entries() : []) {
-      if (index === 0) {
-        let updated = false;
-        for (const inter of stat.data.interfaces ?? []) {
-          if (!updated && selected === 'memory') {
-            updated = true;
-            setSelected(inter.name);
-          }
-          previousRx[inter.name] = inter.counters?.rx_bytes ?? 0;
-          previousTx[inter.name] = inter.counters?.tx_bytes ?? 0;
-        }
-      } else {
-        const newMem = extractMemory(stat.data);
-        memoryData.used.push(newMem.used ?? 0);
-        memoryData.buffered.push(newMem.buffered ?? 0);
-        memoryData.cached.push(newMem.cached ?? 0);
-        memoryData.free.push(newMem.free ?? 0);
-        memoryData.total.push(newMem.total ?? 0);
-        memoryData.recorded.push(stat.recorded);
-
-        for (const inter of stat.data.interfaces ?? []) {
-          const rx = inter.counters?.rx_bytes ?? 0;
-          const tx = inter.counters?.tx_bytes ?? 0;
-          let rxDelta = rx - (previousRx[inter.name] ?? 0);
-          if (rxDelta < 0) rxDelta = 0;
-          let txDelta = tx - (previousTx[inter.name] ?? 0);
-          if (txDelta < 0) txDelta = 0;
-          if (data[inter.name] === undefined)
-            data[inter.name] = {
-              rx: [rxDelta],
-              tx: [txDelta],
-              recorded: [stat.recorded],
-              maxTx: txDelta,
-              maxRx: rxDelta,
-            };
-          else {
-            data[inter.name]?.rx.push(rxDelta);
-            data[inter.name]?.tx.push(txDelta);
-            data[inter.name]?.recorded.push(stat.recorded);
-            // @ts-ignore
-            if (data[inter.name] !== undefined && txDelta > data[inter.name].maxTx) data[inter.name].maxTx = txDelta;
-            // @ts-ignore
-            if (data[inter.name] !== undefined && rxDelta > data[inter.name].maxRx) data[inter.name].maxRx = rxDelta;
-          }
-          previousRx[inter.name] = rx;
-          previousTx[inter.name] = tx;
-        }
-      }
-    }
-
-    return {
-      interfaces: data,
-      memory: memoryData,
-    };
-  }, [getStats.data, getCustomStats.data]);
 
   const interfaces = React.useMemo(() => {
     if (!parsedData) return undefined;
@@ -140,13 +45,6 @@ const DeviceStatisticsCard = ({ serialNumber }: Props) => {
     return <DeviceMemoryChart data={parsedData.memory} />;
   }, [parsedData]);
 
-  const isLoading = React.useMemo(() => {
-    if (!time && getStats?.isFetching) return true;
-    if (time && getCustomStats.isFetching) return true;
-
-    return false;
-  }, [getStats, getCustomStats, time]);
-
   return (
     <Card mb={4}>
       <CardHeader display="block">
@@ -156,7 +54,7 @@ const DeviceStatisticsCard = ({ serialNumber }: Props) => {
           <HStack>
             <ViewLastStatsModal serialNumber={serialNumber} />
             <StatisticsCardDatePickers defaults={time} setTime={setNewTime} onClear={onClear} />
-            <Select value={selected} onChange={onChange}>
+            <Select value={selected} onChange={onSelectInterface}>
               {parsedData?.interfaces
                 ? Object.keys(parsedData.interfaces).map((v) => (
                     <option value={v} key={uuid()}>
@@ -168,9 +66,9 @@ const DeviceStatisticsCard = ({ serialNumber }: Props) => {
             </Select>
             <RefreshButton
               size="sm"
-              onClick={getStats.refetch}
+              onClick={refresh}
               isCompact
-              isFetching={getStats.isFetching}
+              isFetching={isLoading.isLoading}
               // @ts-ignore
               colorScheme="blue"
             />
@@ -188,12 +86,17 @@ const DeviceStatisticsCard = ({ serialNumber }: Props) => {
             </Heading>
           </Flex>
         )}
-        {!getStats?.data && isLoading ? (
-          <Center my="auto">
-            <Spinner size="xl" mt="100px" />
+        {(!parsedData && isLoading.isLoading) || (isLoading.isLoading && isLoading.progress !== undefined) ? (
+          <Center my="auto" mt="100px">
+            {isLoading.progress !== undefined && (
+              <Heading size="md" mr={2}>
+                {isLoading.progress.toFixed(2)}%
+              </Heading>
+            )}
+            <Spinner size="xl" />
           </Center>
         ) : (
-          <LoadingOverlay isLoading={isLoading}>
+          <LoadingOverlay isLoading={isLoading.isLoading}>
             <Box>
               {selected === 'memory' && memory}
               {interfaces}
